// src/app/types/Account.ts
//// React

//// Packages

//// Types
export type SignUpT = {
    username: string
    password: string
}
export type ConfirmSignUpT = {
    username: string
    confirmationcode: string
}
export type ReConfirmSignUpT = {
    username: string
}
export type SignInT = {
    username: string
    password: string
}
export type AuthorizeT = {
    children: React.ReactNode
}

//// Components

//// Middleware & Integrations

//// Application

//// System